requirements.txt
asgiref==3.5.2
Django==3.0.10
djangorestframework==3.13.1
pytz==2022.1
simplejson==3.17.6
sqlparse==0.4.2
